app/requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.1.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
fastapi==0.112.0
fastapi-jwt-auth==0.2.0
fastapi-users==13.0.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pwdlib==0.2.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
websockets==12.0
